rename crazyswarm.py to avoid conflict w/ ros package name
