+++ ros_ws/src/crazyswarm/scripts/pycrazyswarm/__init__.py
@@ -1,3 +1,3 @@
-from .crazyswarm import *
+from .crazyswarm_py import *
 
 __all__ = ["Crazyswarm", "cfsim"]
